requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
python-dateutil==2.8.2
scikit-learn==1.4.2
numpy==1.27.6